pyproject.toml
[tool.poetry]
name = "python-spanner-issue"
version = "0.1.0"
description = ""
authors = ["Mostafa Khalil <mostafa.omar.m.s@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-spanner = "^3.40.1"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
